Re-enable legacy encoding optionally.
The WiFi Alliance WPA3 standard specifies percent encoding in WiFi
URIs. Legacy encoding is to backslash escape the characters \ , ; :
and " only. Commit 3550d86 implemented percent encoding. This commit
re-enables legacy encoding optionally.

The new standard allows all octets to be represented and explicitly
any UTF-8 encoded character can be used in password and ssid
fields. However percent encoding and backslash encoding are
incompatible. For example the string %41 in a passpword would
represent the characters % 4 and 1 with backslash encoding but the
letter A with percent encoding and the string \, would represent the
comma character in backslash encoding but \ and , in percent
encoding. The only way to be compatible with both standards is to restrict
SSIDs and passwords to printable 7 bit ascii excluding the \ , ; : and
" characters.

So make the WPA3 standard the default and enable legacy encoding with
a -L switch.